+++ src/components/Main/Company/Table/Table.styles.jsx
@@ -59,3 +59,4 @@ export const $IconCell = styled.div`
     }
   }
 `
+