+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/14)
 * Create VMenu and Title in Company section with proper styling
+* Create generic endpoint for retrieve entities for manage
+* Create and implement tansack table to view entities
+* Implement sorting in table
 
 0.23 (2024-08-05)
 ===============